src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/element-upgrade-expected.txt

CONSOLE ERROR: line 17: Uncaught TypeError: document.registerElement is not a function
Tests the element upgrade algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
